ipa: rpi: Fix incorrect cast for ExposureTime ControlInfo
controls::ExposureTime is of type ControlTypeInteger32, but the
default ControlInfoMap casts a value to int64_t causing incorrect
initialisation of the associated ControlInfo.

Fix this by casting correctly to int32_t.

Fixes: bea2db5e6190 ("ipa: rpi: Apply default ControlInfo values for sensor controls")

+++ src/ipa/rpi/common/ipa_base.cpp
@@ -62,7 +62,7 @@ const ControlInfoMap::Map ipaControls{
 		      static_cast<int32_t>(controls::ExposureTimeModeManual),
 		      static_cast<int32_t>(controls::ExposureTimeModeAuto)) },
 	{ &controls::ExposureTime,
-	  ControlInfo(1, 66666, static_cast<int64_t>(defaultExposureTime.get<std::micro>())) },
+	  ControlInfo(1, 66666, static_cast<int32_t>(defaultExposureTime.get<std::micro>())) },
 	{ &controls::AnalogueGainMode,
 	  ControlInfo(static_cast<int32_t>(controls::AnalogueGainModeAuto),
 		      static_cast<int32_t>(controls::AnalogueGainModeManual),
